+++ cart/repository/products.json.model.go
@@ -14,7 +14,7 @@ type ProductJSONModel struct {
 	ID          int32  `json:"id"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
-	Amount      int32  `json:"amount"`
+	Amount      int64  `json:"amount"`
 	IsGift      bool   `json:"is_gift"`
 }
